chore: Component filename - Titlecased
